refactor : NetworkStateObserver 추출
